
\name{ldist}
\alias{ldist}
\title{
Compute line length
}
\description{
Computes line length}
\usage{
ldist(p1,p2)
}
\arguments{
  \item{p1}{
(x,y) point on one end of line segment
}
  \item{p2}{
(x,y) point on other end of line segment
}
}

\references{
See Mueller et al., (2016). \url{https://sites.google.com/a/mtu.edu/mapping/}
}
\author{
Shane T. Mueller and Brandon Perelman
}
