

\name{LineMagnitude}
\alias{LineMagnitude}
\title{
Compute line length
}
\description{
Computes line length}
\usage{
LineMagnitude(x1, y1, x2, y2)
}
\arguments{
  \item{x1}{
x coordinate of one end of line
}
  \item{y1}{
y coordinate of one end of line
}
  \item{x2}{
x coordinate of other end of line
}
  \item{y2}{
y coordinate of other end of line
}
}

\references{
See Mueller et al., (under review). \url{https://sites.google.com/a/mtu.edu/mapping/}
}
\author{
Shane T. Mueller and Brandon Perelman
}
