
\name{LLbeta}
\alias{LLbeta}
\title{
Compute Latecki/Lakaemper beta
}
\description{
Compute Latecki/Lakaemper beta
}
\references{
See Mueller et al., (2016). \cr
L. J. Latecki and R. Lakaemper. Convexity Rule for Shape 
Decomposition Based on Discrete Contour Evolution. Computer Vision
and Image Understanding, vol. 73, pp. 441-454, 1999.\cr
\url{https://sites.google.com/a/mtu.edu/mapping/}
}
\author{
Shane T. Mueller and Brandon Perelman
}

