\name{LLKscore}
\alias{LLKscore}
\title{
Compute Latecki/Lakaemper K score
}
\description{
Compute Latecki/Lakaemper K score
}
\references{
See Mueller et al., (under review). \cr
L. J. Latecki and R. Lakaemper. Convexity Rule for Shape Decom-
a position Based on Discrete Contour Evolution. Computer Vision
and Image Understanding, vol. 73, pp. 441-454, 1999.\cr
\url{https://sites.google.com/a/mtu.edu/mapping/}
}
\author{
Shane T. Mueller and Brandon Perelman
}

