\name{PathDist}
\alias{PathDist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute Length of a path
}
\description{
This function computes the length of a path.  It does so by summing each consecutive line segment.
}
\usage{
PathDist(path)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path}{
A two-column matrix describing the x,y coordinates of a path.
}
}
\value{
returns a scalar value indicating path length.
}
\references{
See Mueller et al., (under review). \url{https://sites.google.com/a/mtu.edu/mapping/}
}
\author{
Shane T. Mueller and Brandon Perelman
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


