\name{pathmapping-internal}
\alias{part}
\alias{squarederror}
\title{Internal pathmapping Functions}
\description{
 Internal pathmapping functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }

