\name{pathmapping-package}
\alias{pathmapping-package}
\alias{pathmapping}
\docType{package}
\title{
Determine mapping between paths
}
\description{
This package includes algorithms to find the minimum-cost distance between two paths.
}
\details{
\tabular{ll}{
Package: \tab pathmapping\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-3-4\cr
License: \tab GPL 2.0+\cr
}
Package for computing the least-area mapping between paths.
}

\author{
Shane T. Mueller and Brandon Perelman\cr

Maintainer: Shane T. Mueller <shanem@mtu.edu>
}
\references{
 See Mueller et al., (2015). \cr
\url{https://sites.google.com/a/mtu.edu/mapping/}
\url{https://github.com/stmueller/pathmapping}
}
\keyword{ package }

