\name{pathmapping-package}
\alias{pathmapping-package}
\alias{pathmapping}
\docType{package}
\title{
Compute Deviation and Correspondence Between Spatial Paths
}
\description{
Functions to compute and display the area-based deviation between spatial paths and to compute a mapping based on minimizing area and distance-based cost.  For details, see: "Mueller, S. T., Perelman, B. S., & Veinott, E. S. (2016). An optimization approach for mapping and measuring the divergence and correspondence between paths. Behavior research methods, 48(1), 53-71."
}
\details{
\tabular{ll}{
Package: \tab pathmapping\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-03-15\cr
License: \tab GPL 2.0\cr
}
Computing the least-area mapping between paths.
}

\author{
Shane T. Mueller and Brandon Perelman\cr

Maintainer: Shane T. Mueller <shanem@mtu.edu>
}
\references{
 See Mueller et al., (2016). \cr
\url{https://sites.google.com/a/mtu.edu/mapping/}
}
\keyword{ package }

